% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enumerateG_revised-function.R
\name{enumerateG_revised}
\alias{enumerateG_revised}
\title{Enumerate genotypic vectors}
\usage{
enumerateG_revised(k, S.list)
}
\arguments{
\item{k}{(numeric integer) specifies the number of putatitive replicates associated with
              the proband in question.}

\item{S.list}{(list) output from \code{\link{enumerateS}}.}
}
\value{
The function returns a list of length two:
     (1) `G|S`: (list) enumerates all genotype vectors consistent with source vector s
     (2) `P(G|S)`: (vector) provides the probability of genotype vector g given source vector s.
                   We compute P(G|S) assuming the uniform probability law (i.e. 1/cardinality of set).
}
\description{
Enumerates all genotypic vectors of length k consistent with each identity vector.
}
\author{
Ariel W Chan, \email{ac2278@cornell.edu}
}

