% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enumerateI-function.R
\name{enumerateI}
\alias{enumerateI}
\title{Enumerate identity vectors}
\usage{
enumerateI(k)
}
\arguments{
\item{k}{(numeric integer) specifies the number of putative replicates associated with
       the proband in question.}
}
\value{
The function returns a named list enumerating all of the identity vectors that describe how
        the k putative replicates could potentially be related to one another.
}
\description{
Enumerates all identity vectors associated with the k putative replicates of a given proband
}
\examples{
k3 <- enumerateI(k=3)
This function call returns a named list consisting of three elements, which are also themselves
lists: 's=1', 's=2', and 's=3'.

Printing the output of enumerateI(k=3)[['s=2']] allows us to see the identity vectors that describe the case
where the k=3 supposed replicates actually originated from s=2 genotypic sources:
$`120`
[,1] [,2] [,3]
[1,]    1    2    2     # corresponds to the identity vector I=1,2,2, where samples d=2 and d=3 originate from one genotype source and d=1 is assumed to be the outlier
$`210`
[,1] [,2] [,3]
[1,]    1    1    2     # corresponds to the identity vector I=1,1,2, where samples d=1 and d=2 originate from one genotype source and d=3 is assumed to be the outlier
[2,]    1    2    1     # corresponds to the identity vector I=1,2,1, where samples d=1 and d=3 originate from one genotype source and d=2 is assumed to be the outlier
}
\author{
Ariel W Chan, \email{ac2278@cornell.edu}
}

