% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enumerateS-function.R
\name{enumerateS}
\alias{enumerateS}
\title{Enumerate source vectors}
\usage{
enumerateS(k)
}
\arguments{
\item{k}{(numeric integer) specifies the number of putative replicates associated with
       the proband in question.}
}
\value{
The function returns a named list enumerating all of the source vectors that describe how
        the k putative replicates could potentially be related to one another.
}
\description{
Enumerates all source vectors associated with the k putative replicates of a given proband
}
\examples{
k3 <- enumerateS(k=3)
This function call returns a named list consisting of three elements, which are also themselves
lists: 'i=1', 'i=2', and 'i=3'.

Printing the output of enumerateS(k=3)[['i=2']] allows us to see the source vectors that describe the case
where the k=3 supposed replicates actually originated from i=2 genotypic sources:
$`120`
[,1] [,2] [,3]
[1,]    1    2    2     # corresponds to the source vector S=1,2,2, where samples d=2 and d=3 originate from one genotype source and d=1 is assumed to be the outlier
$`210`
[,1] [,2] [,3]
[1,]    1    1    2     # corresponds to the source vector S=1,1,2, where samples d=1 and d=2 originate from one genotype source and d=3 is assumed to be the outlier
[2,]    1    2    1     # corresponds to the source vector S=1,2,1, where samples d=1 and d=3 originate from one genotype source and d=2 is assumed to be the outlier
}
\author{
Ariel W Chan, \email{ac2278@cornell.edu}
}

