% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/thin-function.R
\name{thin}
\alias{thin}
\title{Thin sites}
\usage{
thin(x, distance)
}
\arguments{
\item{x}{(numeric vector) specifies the set of sites the user wishes to thin.}

\item{distance}{(numeric) specifies the desired minimum distance (in bp) between any two sites in the
                     resulting set}
}
\value{
The function returns a set of thinned sites where no two sites are within a specified distance
        (in bp) of each other.
}
\description{
Thins a set of sites such that no two sites are within a specified distance (in bp) from one another.
}
\author{
Ariel W Chan, \email{ac2278@cornell.edu}
}

